DVB Refactoring, removes ES_Type.
ES_Type was a processed form of the PMT type, it is one of the things preventing us from using pmt directly from the signal monitor when we start up the DVB Recorder. Eliminating it plus some other changes should allow us to shave a couple hundred milliseconds off the DVB tuning time.



git-svn-id: http://svn.mythtv.org/trunk@9085 4b91aca1-2a1f-0410-a73a-a78f2df169e8