References #1403. Fixes segfault.
The assert was being triggered because the caption packet buffer grew past a certain point which it shouldn't. This replaces the assert with a realloc.

But this shouldn't be happening in the first place. 

teagardn, can you change these three lines in cc708decoder.cpp:
{{{
#define DEBUG_CC_SERVICE_2     0
#define DEBUG_CC_RAWPACKET     0
#define DEBUG_CC_SERVICE_BLOCK 0
}}}
to
{{{
#define DEBUG_CC_SERVICE_2     1
#define DEBUG_CC_RAWPACKET     1
#define DEBUG_CC_SERVICE_BLOCK 1
}}}

Then send attach the output from the watcing the same file to about 20 seconds past the point where it crashed before this commit.



git-svn-id: http://svn.mythtv.org/trunk@9166 4b91aca1-2a1f-0410-a73a-a78f2df169e8