Set the ~realtime flagging 'required Buffer' (lag-behind) time to 3 seconds, down from 30.  Bruce Markey did some testing on this and it appears we can go at least this low without incident.
Also modify the 'requiredHeadStart' calculation to be based on this 3-second
lag-behind time instead requiring at least 60 seconds.  When performing logo
detection, the requiredHeadStart is still adjusted to take the preroll and
the required frames for logo detection into account.

Add another case to the speed-up/slow-down calculation for the ~realtime
flagger to include a middle-ground for when we are the exact 'requiredBuffer'
seconds behind and tweak the speed-up code a bit to allow the flagger to
catch up to realtime a little faster.



git-svn-id: http://svn.mythtv.org/trunk@9237 4b91aca1-2a1f-0410-a73a-a78f2df169e8