Closes #1433, by applying patch.
This tweaks some params in the pes memory allocator to get a big bump in performace.

The method causing most of the allocations will actually be a zero memory copy routine when the SIParser refactor is done, but these optimizations will still be useful for when we need to clone tables.



git-svn-id: http://svn.mythtv.org/trunk@9238 4b91aca1-2a1f-0410-a73a-a78f2df169e8