References #1095. UI cleanup for captions/subtitle/teletext.
Puts the EIA-608 and EIA-708 captions in the track selection system previously only used for subtitles and audio. This also tells you what language these streams are in if this information is available in the PMT.

Puts all the caption types into an SetCaptionsEnabled(bool) method which selects the best set available. TOGGLECC uses this function to toggle the best available caption. For instance, if two captions in a different language are available the one in your most preferred language is chosen, subtitles are preferred over teletext, etc.

Adds TOGGLE bindings for each type of caption and uses that in the individual caption menus rather than the global TOGGLECC.

Adds a NEXTCC binding which allows you to go through all the available caption types including teletext menus in PAL countries.

None of the new bindings have a keyboard binding by default, and I've removed the "^" default keybinding for TOGGLECC708. NEXTCC is the only one I could see as being of general use.



git-svn-id: http://svn.mythtv.org/trunk@9291 4b91aca1-2a1f-0410-a73a-a78f2df169e8