Fixes #744. Switch over to DVBSignalMonitor based channel scan from SIParser based.
This may cause some problems in DVB land with the channel scan.

Please attach a "mythtv-setup -v siparser" log to any ticket...

This will allow us to greatly simplify the scanner and SIParser, but I have left in the SIParser hooks for now. Simply uncomment the "#define USE_SIPARSER" in siscan.h to switch back. But that code will be removed once I'm confident that most of the bugs are gone from the newer scanning code, so please report any problems instead of just uncommenting the define.



git-svn-id: http://svn.mythtv.org/trunk@9305 4b91aca1-2a1f-0410-a73a-a78f2df169e8