References #1095. Reorder ATSC caption packets properly.
The caption packets are in display order, but we were processing them in decode order. Many channels transmit captions only on B frames, so this wasn't a problem on those channels. But for channels like FOX and UPN to work we need to do this properly.

This also makes the implementation closer to what the ffmpeg folks want for ATSC caption decoding to be accepted upstream, by using a stream of data rather than a callback; it just needs to be repackaged into an av_stream before they will accept the patch.



git-svn-id: http://svn.mythtv.org/trunk@9312 4b91aca1-2a1f-0410-a73a-a78f2df169e8