References #1341. Portion of DSMCC/MHEG patch from David Matthews.
This adds the DSMCC stream type and parses out some info for it out of the PMT table.

This also fixes a long standing problem with the mpegts parser not understanding multiple PES packets in a TS packets. MythTV does not generate such streams, but this fix allows you to watch some raw stream captures in mythtv which mplayer can't decode.



git-svn-id: http://svn.mythtv.org/trunk@9314 4b91aca1-2a1f-0410-a73a-a78f2df169e8