Don't update the recgroup on existing recordings when changing their scheduled recording entry.  This closes #1505.
I don't think that we need to be updating the recgroup on existing recordings
since the recgroup for an existing recording can be editted, and groups of
recordings can have their recgroup changed using the 'playlist' functionality
on the Watch Recordings screen.  There are 3 other fields which are updated
here as well: recpriority, transcoder, and playgroup.  The playgroup is
another iffy item, since we can edit that from the Watch Recordings screen,
but I am not changing that in this commit (for now).  The recpriority and
transcoder fields are not edittable anywhere else so those are left alone
as well.



git-svn-id: http://svn.mythtv.org/trunk@9392 4b91aca1-2a1f-0410-a73a-a78f2df169e8