Adds some more XDS decoding to CCDecoder.
In particular this decodes more of the V-Chip data and the cable network name and network call letters.

There are still some problems with XDS decoding, most importantly it doesn't cope properly with breaks in the XDS data and it doesn't checksum the XDS packets. I don't have the spec for XDS, but if someone does fixing these would make XDS data much more reliable.



git-svn-id: http://svn.mythtv.org/trunk@9416 4b91aca1-2a1f-0410-a73a-a78f2df169e8