References #1485. Checks each param to dtv_multiplex get/insert for validity before use.
This should fix some issues with uninitialized values when inserting a new multiplex.

This also fixes a problem with the CreateMultiplex() functions created by [9537]. Because we did not always insert a multiplex when requested in insert_dtv_multiplex() the function that got the multiplex to return from CreateMultiplex() returned the wrong multiplex. We now return the multiplex from insert_dtv_multiplex so that CreateMultiplex() has something valid to return.

Finally, mythfilldatabase.pro did not have the libmythtv in it's depend path so changes to channelutil.h caused linking errors, this has been fixed.



git-svn-id: http://svn.mythtv.org/trunk@9539 4b91aca1-2a1f-0410-a73a-a78f2df169e8