Some more DataDirect refactoring, improvements.
This uses some of the functions added in [9621] to get the right data for channel editing in tv_play.cpp

This also changes some of the container classes in DataDirectProcessor to make things more efficient + other refactors.

Finally, this speeds up the "Fetch channels from listings source" button in mythtv-setup by only fetching listings for one day and only on the specific sourceid we're interested in.



git-svn-id: http://svn.mythtv.org/trunk@9625 4b91aca1-2a1f-0410-a73a-a78f2df169e8