Adds a new param to mythfilldatabase "--only-update-channels".
This is used now where we only need to get channels, and do not need any listings data. With datadirect this means we don't get any listings data, with other grabbers we grab one day of data. I believe at least some of the scripts already support getting just the channels, but I'll leave it to the people who use those scripts to make patches...

This also avoids running the channel update for every day of the listings, and only runs it for the first day of listings.



git-svn-id: http://svn.mythtv.org/trunk@9632 4b91aca1-2a1f-0410-a73a-a78f2df169e8