Fixes #1337. By adding a recording profile option that allows you to require video on the channel for it to be "On-Air".
By default the behavior doesn't change and all streams are recorded and none are required in DVB land.

However there are two additional recording types: "TV Only" and "Audio Only". If "TV Only" is selected only channels with video are considered "On-Air", and only the audio, video and captioning streams are recorded even if other streams are present in the program. If "Audio Only" is selected, only the audio is required, but all streams in the program are recorded.



git-svn-id: http://svn.mythtv.org/trunk@9638 4b91aca1-2a1f-0410-a73a-a78f2df169e8