This fixes some of the spacing problems using the *ConfigurationGroup's + adds a new as yet unused ConfigurationWizard.
I was working on a configuration wizard that would allow you to jump to a specific configuration pane when I realized there was a problem with how we initialize the ConfigurationGroup variables when inheriting from VerticalConfigurationGroup.

Since the ConfigurationGroup is virtually inherited in the VerticalConfigurationGroup we need to specifically initialize the ConfigurationGroup parent when inheriting from the VerticalConfigurationGroup, but very few classes were doing this, and instead were inheriting the default settings. Many of these classes then went ahead and reinitialized these variables to get rid of frames & labels.

I've also added a JumpConfigurationWizard, if you want to try it now, just replace ConfigurationWizard in globalsettings.h with JumpConfigurationWizard. But at the moment it uses buttons and is a bit ugly and klunky. The plan is to use a combobox to pick a pane to jump to, but this may requires some reorganization of the configuration panes which I'm still looking at.

We should not be virtually inheriting nor using multiple inheritence with classes that use Qt signals and slots, but I think this can only be fixed with a serious rewrite of the configuration code.



git-svn-id: http://svn.mythtv.org/trunk@9646 4b91aca1-2a1f-0410-a73a-a78f2df169e8