Closes #1147. Adds limit to the number of times we try to generate the preview image.
The core problem in the ticket, segfault when permissions are broken, has already been
fixed in svn. This fixes the other problem, which is that we try to regenerate the
preview continously, even if something keeps preventing us from completing the 
procedure. This limits the number of retries to five, which should capture most of
the real cases we should recreate the preview.



git-svn-id: http://svn.mythtv.org/trunk@9691 4b91aca1-2a1f-0410-a73a-a78f2df169e8