Closes #850. Fixes rounding error in timecode calculation for streams without an explicit timecode.
This applies the patch from mythtv At dadeos DT freeserve dot co dt uk, but without the changes to
the signature of getAudiotime(). The current return type of int can handle 500 hours of audio
without wackyness, and if we are going to change the binary revision we should fix the constness
and return types of the rest of the methods in the audio classes.



git-svn-id: http://svn.mythtv.org/trunk@9770 4b91aca1-2a1f-0410-a73a-a78f2df169e8