Refs #1485. Makes multiplex updates smarter.
This avoids inserting redundant multiplexes by updating existing ones when possible and not requiring an exact match on some of the optional params (i.e. if we has 'auto' before but now know the exact value we don't insert a new multiplex, we just update the existing one.

This has been tested with DVB, ATSC & raw MPEG transports.



git-svn-id: http://svn.mythtv.org/trunk@9822 4b91aca1-2a1f-0410-a73a-a78f2df169e8