Refs #1485. This increases the timeout when scanning for DVB tables.
The NIT table is only required to be sent once every 10 seconds.
I've added 22 seconds to the time we spend on each transport when
scanning using the hardware used in Europe and Australia. This
gives us two opportunities to see any single NIT table.

I've also added a ModulationDB() method to the transport scan item,
so that the correct modulation can be inserted for multiplexes when
scanning using the HDTV or HDHomeRun recorders and MythTV is not
compiled with DVB driver support. Before, we always inserted '8vsb',
but the HDHomeRun device should support QAM soon.



git-svn-id: http://svn.mythtv.org/trunk@9839 4b91aca1-2a1f-0410-a73a-a78f2df169e8