Refs #1035. Big EIT branch merge.
This moves the EIT handling out of SIParser and mostly into the DVB and ATSC stream handlers, allowing recorders other than the DVBRecorder to handle EIT without massive code duplication. This also allows us to get rid of a bunch of dependencies on the dvbdev stuff. The only thing we are still using there is the CAM code.

This also fixes several problems in the MPEG PES packet assembly code that were uncovered when we started using it for EIT in addition to using it for the tuning tables.

Since SIParser is gone this also allows us to use the table caching in the stream handler code to speed up intra-transport tuning. If we already have the needed tuning tables we skip the signal monitoring step and begin recording the new channel immediately.

This also fixes various bugs in EIT handling, and lowers the EIT processors memory consumtion further.

This also removes the arbitrary limit of four DVB cards per backend. Now you should be able to add as many DVB devices as your hardware allows.

I recommend a "make distclean".



git-svn-id: http://svn.mythtv.org/trunk@9957 4b91aca1-2a1f-0410-a73a-a78f2df169e8