Fixes #1485. Refs #1704. Adds an MPEG1/MPEG2/MPEG4-AVC IP network recorder.
This is another contribution from the Queen of Canada and the good folks at CRC.

There are instructions in crcipnetworkrecorder.cpp on how to broadcast your own channel using vlc.

The MPEG4-AVC keyframe detection code is in DTVRecorder so it could be used by the DVBRecorder to generate a correct keyframe map for the UK HDTV trial streams.



git-svn-id: http://svn.mythtv.org/trunk@9966 4b91aca1-2a1f-0410-a73a-a78f2df169e8