Fixes #1567. Adjust for non-standard locations where Debian sticks the nvidia's opengl headers.
Since nVidia's driver packages are not considered part of the distribution they aren't installed in the proper locations, but instead are put in documentation or share directories from which the user is expected to manually install them. Since few users do this, this just makes MythTV check those locations and adds them to the include path if needed.



git-svn-id: http://svn.mythtv.org/trunk@9977 4b91aca1-2a1f-0410-a73a-a78f2df169e8