Make the DVD menu music and length themeable.
Fix the chapter menu music not looping.

Move the menu music files and some common images into their own directory
so they can be shared by all themes.

Change the theme file lookup so it first looks in the theme directory 
then if not found looks in the shared music and image directories.

Don't run a TS file though mythtranscode if it has to be re-encoded anyway.

git-svn-id: http://svn.mythtv.org/trunk@9987 4b91aca1-2a1f-0410-a73a-a78f2df169e8