Refs #1104. Apply soundtouch portion of multi channel audio support.
You need to use different calls to use this code, so just changing the audio channel define in AvFormatDecoder will not enable this. I haven't made the call changes because the MMX optimized version of one of the multi-channel soundtouch functions is not complete.

But you can test this using the mythtv_ac3.17.patch on the #1104 ticket.

The changes are not extensive. The buffer needs to be be bigger to fit the extra audio channels, and the audio processing functions need to use a bigger stride.



git-svn-id: http://svn.mythtv.org/trunk@9998 4b91aca1-2a1f-0410-a73a-a78f2df169e8