Fixes #1847. Make automatic input switching (based on channel number) work with DVB cards w/patch from Mark Buechler.
When we made DVBChannel input aware a few months we forgot aspect of this required so that you could switch to a channel on another input without explicitly switch inputs using the "C" key first. It actually worked as a side effect of the full retuning we were doing on each DTV channel change, but this broke when I added the quick intra-multiplex channel changing code.



git-svn-id: http://svn.mythtv.org/trunk@10023 4b91aca1-2a1f-0410-a73a-a78f2df169e8