Closes #303. Enhances media detection by looking at file name extensions with modified patch from Leo Weppelman.
If we can't detect a type other than "DATA" by any other means (such as used for DVDs and CDs), then with this patch we look at the file extensions, and select the media handler which will handle the greatest number of files on the device.

For example we count things such as "gif,jpg,png" and "ogg,mp3,aac,flac" and use the one with the highest file count to select mythgallery or mythmusic, resp. (This is assuming both media plugins are present.)

BTW This is surprisingly fast for the photo cd's I have laying around. 


NOTE:  This changes the MYTH_BINARY_VERSION, so plugins must be recompiled and installed after you've installed the updated mythtv libs.




git-svn-id: http://svn.mythtv.org/trunk@10095 4b91aca1-2a1f-0410-a73a-a78f2df169e8