Refs #1887. Fixes the underlaying cause of the invalid PAT in the DVB recordings which causes the reported crash.
Multiple PAT tables do not seem to be precluded in the MPEG-2 spec, but seem to have been banned in DVB-S, DVB-T, DVB-C and ATSC. But they are now allowed in DVB-S2, and have been seen in the wild.

Multiple PAT sections have always been allowed by the spec, but we haven't seen them in the wild.

This extends mpegstreamdata to allow both multiple sections per PAT table and multiple PAT tables. This changes the interface a bit, hence the changes in the other files.

I've left the ticket open because still haven't looked at having ffmpeg not crash on bad tables.



git-svn-id: http://svn.mythtv.org/trunk@10097 4b91aca1-2a1f-0410-a73a-a78f2df169e8