Added a "Store" function in the Custom Record editor. This will save the current name and clauses in the 'customexample' table. These are then added to the end of the list in the examples ComboBox marked as "<name> (stored example)".
This serves several purposes:

- If there is a type of rule you like or a trick you often use, you
can create a template and Store it. You will then be able to Add your
example just as you would Add any of the built-in examples.

- If there is a complex query you would like to use for searching but
do not want to save it as a recording rule, it can be stored then
recalled in the future when you want to run your search again.

- If a new rule is problematic, you can store your work in progress,
read docs, ponder then come back to it later by recalling the stored
text to continue working.

If the name matches an existing stored example, the button on the
dialog page will say "Replace" and it will overwrite the existing
entry.

If the example ComboBox is showing a "(stored example)" when the
"Store" button is clicked, the dialog page will also show a "Delete"
button for that stored example.


git-svn-id: http://svn.mythtv.org/trunk@10131 4b91aca1-2a1f-0410-a73a-a78f2df169e8