Added per rule preferred input plus other scheduling and priority options.
The "Scheduling Options" section now has a widget to select a
preferred input for any rule. The default is "Use any available
input" with the other choices matching the strings that appears in
the OSD when you change inputs. IOW displayname else <cardid>:
<inputname>. If an input is selected and is available during the
placement phase of the scheduler run, showings on the preferred
input will be considered ahead of the showings on other inputs. If
the input is unavailable or is already assigned to another higher
priority show, another showing will be chosen, This has several
benefits and variations of these examples may be useful for you.

My card 2 is on my master in the living room and it is more
convenient to record sporting events on this card. In the past I
would give games low priority so they would lose to other shows
that would win card 1 forcing the game to card 2. Now I can give
games high priority (because I really want to see them =) and say
that they should grab card 2 as the best first choice.

My card 1 is on a slave where the reception is better on a few high
and low numbered analog channels, OLN on 73 for example. OLN had
several showings per week for the Iditarod where I set low priority
rule to prefer card 1. The first showing at 7PM would lose to a higher
priority show. Rather than choosing the 7PM showing on card 2 with
poor reception, it would choose a later showing on card 1 with a good
signal.

I get NBC on channel 3 and digital cable 123. While the digital
version may be better, this channel has public service programming
during The Tonight Show and Ebert & Roeper. I used to use channel
priorities to switch which channel was preferred throughout the week.
Now I leave channel 3 as the default choice and set individual titles
like "The Apprentice" to prefer the digital cable input.

A couple years ago, someone built two identical backend servers with
local disks only to find that the majority of shows were recorded
on the host with card 1. With preferred input, a few titles could
be 'assigned' to the cards on the other host to load balance the
disk usage.

My card 4 is on a host where I only run a backend part time when
needed.  This has a lower input priority so that it is only used
when necessary.  I have a custom rule for "Breaking News" which
records every show on CNN up to a cutoff time when there is a major
news story. I set this rule's preferred input to 4 so that it will
use the temporary slave for 24 hour news and leave the primary cards
for regular programming. If the backend for card 4 is not running,
"Breaking News" shows would be scheduled along with the other shows
on the available cards as usual.

And, of course, you can add an override for one showing from a series
to specify which input to use in order to solve a unique, one time
problem.

The amount of priority boost for preferred inputs can be adjusted
in TV Settings->Recording Priorities. This page includes other new
options. 

Added the simple, obvious and long overdue DefaultStartOffset and
DefaultEndOffset. These will set the initial 'Start Early' and 'End
Late' values when a new record rule is created. You can tweak each
rule as needed while you are creating the rule and modify any rule
at any later time. These defaults are only applied as the attributes
of a brand new rule are being initialized. This is beneficial for TV
markets where content extends minutes outside of the times in the TV
listings for the majority of shows. Adding this time will assure
that the highest priority shows will record to completion. Lower
priority shows may result in conflicts and can be resolved for the
user's preference on a case by case basis.  Because each station and
timeslot for shows vary, rules can be tweaked individually without
affecting the extra time for any other shows (i.e. one show always
ends on time, another is usually 10 to 12 minutes late. etc.).

In recent months the priority scheme had quietly changed and is
undocumented. This is a regression from the original concept
described in the mythtv-HOWTO.  The "Complex Prioritization" option
would use this recent scheme whereas the default is to use the more
obvious total priority approach.

Priority can be added for showings marked as Closed Captioned. This
was suggested by a hearing impaired myth user in England. At that time,
his grabber was not marking shows as program.closecaptioned and I
hope that this has been or will be fixed. However, users of DataDirect
and possibly other grabbers will be able to prefer showings with CC
over those without.

Priority can be added for showings marked as HDTV. This doesn't say 
for a fact that the broadcast actually is or isn't HDTV or if the input
has an HDTV device. The priority is added if the listings say that the
showing is marked as HDTV and is only as useful as the data is accurate.


git-svn-id: http://svn.mythtv.org/trunk@10186 4b91aca1-2a1f-0410-a73a-a78f2df169e8